+++ src/resolvers/Mutation/user.js
@@ -1,27 +1,34 @@
 const { hashPassword } = require("../../utils/index");
+const { generateJWT } = require("../../utils/jwt");
 
 const signup = async (
   parent,
   { input: { firstName, lastName, userName, emailAddress, password } },
-  { models }
+  { models, req }
 ) => {
   const hash = hashPassword(password);
+  const trimEmail = emailAddress.trim().toLowerCase();
   try {
     const saveData = await models.User.create({
       firstName,
       lastName,
       userName,
-      emailAddress,
+      emailAddress: trimEmail,
       password: hash,
     });
 
-    return saveData;
+    return {
+      token: generateJWT(saveData),
+    };
   } catch (err) {
     console.error(err);
   }
 };
 
-const deleteUser = async (parent, { id }, { models }) => {
+const deleteUser = async (parent, { id }, { models, req }) => {
+  if (!req.id) {
+    throw new Error("You must be signed in");
+  }
   try {
     return await models.User.findByIdAndDelete(id);
   } catch (err) {
@@ -45,7 +52,7 @@ const updateUser = async (
       dob,
     },
   },
-  { models }
+  { models, req }
 ) => {
   try {
     const updateData = await models.User.findByIdAndUpdate(
@@ -73,7 +80,10 @@ const updateUser = async (
   }
 };
 
-const verifyUser = async (parent, { id }, { models }) => {
+const verifyUser = async (parent, { id }, { models, req }) => {
+  if (!req.id) {
+    throw new Error("You must be signed in");
+  }
   try {
     const updateData = await models.User.findByIdAndUpdate(id, {
       verified: true,
